Refactor npm2rpm
This is a pretty big rewrite to address various issues around package
naming. NPM allows packages to be named as @group/package which creates
issues when mapping to RPMs since we want to avoid slashes. They create
directories on the filesystem and appears to break some assumptions.

To address this we normalize package names (see getRpmPackageName). We
also stop downloading dependencies altogether. The user is now required
to do this themselves. The suggested solution is:

spectool -g nodejs-*.spec.

We also made the output directory configurable. To do this we did
flatten the structure. This makes it easier to integrate into the
packaging structure we use in foreman-packaging.

The big downside is that if two dependencies have the same name but in a
different namespace, we can't support this because the resulting
filename is the same. This is currently a known limitation.

It also moves from var to const and string interpolation to create a
more modern javascript style.