Fix the handling of semver ranges that contain '*'
Use the semver library to parse version strings into ranges. For any
version strings that fail to parse or are returned as '*', output a
"Requires:" line with no version information.

Also skip the version restriction for ranges that include '||'
alternatives. These could be handled in the future using RPM's boolean
dependency syntax.