Purge the NPM registry output from duplicate entries
This is based on what I wrote[1]. The short summary is NPM offline
installs are achieved by having a cache of responses from the NPM
registry. This cache was already primed with tarballs in the build
process (using npm cache add in the RPM spec), but the tarball contained
the same duplicated files.

By priming the cache in the same way and excluding the duplicated
entries space can be saved. For @theforeman/builder 10.1.7 the
compressed responses are reduced from 6.7M to 4.6M, about 31%.

[1]: https://community.theforeman.org/t/trimming-foreman-packaging-branches/29189/7